refactor: unified SettingCard component with Vercel-style pattern
- Create shared SettingCard component in src/components/
- Split service settings into individual cards (health check, timeouts, limits)
- Split project settings into individual cards (name, delete)
- Split cleanup settings into individual cards (auto cleanup, images, prune)
- Update session section with sign out in footer
- Always show form (no edit mode toggle)
- Each card has own Save button in footer
- Add JSDoc documentation with usage guidelines
