+++ src/app/settings/_components/github-section.tsx
@@ -12,8 +12,8 @@ import {
 import Link from "next/link";
 import { useSearchParams } from "next/navigation";
 import { useEffect, useRef, useState } from "react";
+import { SettingCard } from "@/components/setting-card";
 import { Button } from "@/components/ui/button";
-import { SettingCard } from "./setting-card";
 
 interface GitHubInstallation {
   id: string;
