+++ src/app/settings/_components/session-section.tsx
@@ -1,7 +1,7 @@
 "use client";
 
+import { SettingCard } from "@/components/setting-card";
 import { Button } from "@/components/ui/button";
-import { SettingCard } from "./setting-card";
 
 export function SessionSection() {
   async function handleSignOut() {
@@ -10,10 +10,19 @@ export function SessionSection() {
   }
 
   return (
-    <SettingCard title="Session" description="Manage your current session.">
-      <Button variant="outline" onClick={handleSignOut}>
-        Sign out
-      </Button>
+    <SettingCard
+      title="Session"
+      description="You are currently signed in. Sign out to end your session and return to the login page."
+      footerRight={
+        <Button variant="outline" onClick={handleSignOut}>
+          Sign out
+        </Button>
+      }
+    >
+      <p className="text-sm text-neutral-400">
+        Signing out will clear your session. You'll need to enter the install
+        password to sign back in.
+      </p>
     </SettingCard>
   );
 }
