+++ src/app/settings/_components/ssl-section.tsx
@@ -2,10 +2,10 @@
 
 import { CheckCircle2, Loader2, XCircle } from "lucide-react";
 import { useEffect, useState } from "react";
+import { SettingCard } from "@/components/setting-card";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { SettingCard } from "./setting-card";
 
 interface DnsStatus {
   valid: boolean;
